Extract importable package from CLI
The motivation here is to remove the need for forking prealloc for use
in linter aggregation tools such as golangci-lint. I attempted to make
as few changes as possible to the code itself to keep the existing
structure, while adding the interface necessary for golangci-lint to use
it.

Those changes are based largely on the commit here:
https://github.com/golangci/prealloc/commit/215b22d4de21190b80ce05e7d8466677c1aa3223
