+++ go.mod
@@ -1,0 +1,3 @@
+module github.com/alexkohler/prealloc
+
+go 1.15
